pyproject.toml
[tool.poetry]
name = "cam2 Image Depth Maps"
version = "0.0.1"
description = ""
license = ""
authors = ["Nicholas M. Synovic <nicholas.synovic@gmail.com>"]
maintainers = ["Nicholas M. Synovic <nicholas.synovic@gmail.com>"]
readme = "README.md"
homepage = ""
documentation = ""
keywords = []
classifiers = []

[tool.poetry.dependencies]
matplotlib = "^3.4.2"
numpy = "^1.20.3"
opencv_python = "^4.5.3.56"
pandas = "^1.2.5"
Pillow = "^8.4.0"
python = "^3.10"
torch = "^1.11.0"

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
bandit = "^1.7.4"
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.950"
pylint = "^2.13.9"
pytest = "^5.2"
lxml = "^4.8.0"

[tool.poetry.scripts]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/NicholasSynovic/cam2_ImageDepthMaps/issues"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
